Time: 0 ms (100.00%) | Memory: 44.7 MB (38.13%) - LeetSync

+++ 3251-maximum-area-of-longest-diagonal-rectangle/maximum-area-of-longest-diagonal-rectangle.java
@@ -1,16 +1,15 @@
 class Solution {
     public int areaOfMaxDiagonal(int[][] dimensions) {
-        double diagonal = 0;
+        int diagonal = 0;
         int area = 0;
         for(int[] sides: dimensions){
             int l = sides[0];
             int b = sides[1];
-            double newDiag = l*l + b*b;
+            int newDiag = l*l + b*b;
             if(diagonal < newDiag){
                 diagonal = newDiag;
                 area = l*b;
             }
-
             else if(diagonal == newDiag){
                 area = area < l*b ? l*b:area; 
             }
